Export the extension as the main module
  - Makes extension loader a litt nicer
    ("buster-lint" vs "buster-lint#extension")
  - Package is 'only' a buster extension, no need to export anything else
    (use autolint directly for other linting needs)
