+++ lib/buster-lint.js
@@ -1,4 +1,19 @@
+var lint = require("./lint");
+
 module.exports = {
-    lint: require("./lint"),
-    extension: require("./extension")
+    create: function (options) {
+        var instance = Object.create(this);
+        instance.checker = lint.create(options || {});
+        return instance;
+    },
+
+    beforeRun: function (resourceSet, analyzer) {
+        resourceSet.addProcessor(function (resource, content) {
+            var result = this.checker.check(content, resource.path);
+            if (!result.ok) {
+                var path = resource.path;
+                analyzer.error("Lint in " + path, result);
+            }
+        }.bind(this));
+    }
 };
